+++ client/gldt_swap_app/src/components/swap/reverse/from/index.tsx
@@ -3,33 +3,42 @@ import { useNft } from "@context/index";
 import Card from "@components/shared/card/Base";
 
 const ReverseSwapFrom = () => {
-  const { getSelectedTotalGLDT, getSelectedTotal } = useNft();
-  const totalNFTs = getSelectedTotal();
-  const totalGLDTtoSwap = getSelectedTotalGLDT();
+  const {
+    getSelectedTotalGLDT,
+    getSelectedTotalNFTs,
+    getSelectedTotalGLDTWithFees,
+  } = useNft();
+  const totalNFTs = getSelectedTotalNFTs();
+  const totalGLDT = getSelectedTotalGLDT();
+  const totalGLDTWithFees = getSelectedTotalGLDTWithFees();
 
   return (
     <div className="">
       <Card>
         <div className="flex justify-center items-center">
           <div className="font-semibold text-content/40">
-            {totalGLDTtoSwap + totalNFTs} GLDT
+            {totalGLDTWithFees.string} GLDT
           </div>
         </div>
       </Card>
-      {totalNFTs > 0 && (
+
+      {totalNFTs.number > 0 && (
         <div className="mt-4 text-sm px-4">
-          As you are swapping for{" "}
-          <span className="font-semibold text-gold">{totalNFTs} GLD NFTs</span>,
-          you are charged{" "}
-          <span className="font-semibold text-gold">{totalNFTs} x 1 GLDT</span>{" "}
-          for the reverse swap.
-          <br />
+          You are burning{" "}
           <span className="font-semibold text-gold">
-            {totalGLDTtoSwap} GLDT
+            {totalGLDT.string} GLDT
           </span>{" "}
-          will be burned and{" "}
-          <span className="font-semibold text-gold">{totalNFTs} GLDT</span> go
-          to Gold DAO treasury.
+          and will receive{" "}
+          <span className="font-semibold text-gold">
+            {totalNFTs.string} GLD NFTs
+          </span>
+          .
+          <br />
+          For this, you are charged a fee of{" "}
+          <span className="font-semibold text-gold">
+            {totalNFTs.string} x 1 GLDT
+          </span>{" "}
+          which will go to the Gold DAO treasury.
         </div>
       )}
     </div>