+++ client/gldt_swap_app/src/components/swap/reverse/insufficient-gldt-disclaimer/index.tsx
@@ -1,23 +1,20 @@
-import { useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
 import { Button } from "@components/ui";
 
 import { useReverseSwapProceed } from "@context/index";
 
 const InsufficientGLDTDisclaimer = ({ className }: { className?: string }) => {
-  const navigate = useNavigate();
   const { state } = useReverseSwapProceed();
   const { totalSwapGLDT, balanceGLDT, countSelectedNfts } = state;
-
-  const handleClickGetGLDT = () => {
-    navigate("/swap?view=0&mode=0");
-  };
+  const linkIcpSwapPairICPGLDT =
+    "https://app.icpswap.com/swap?input=ryjl3-tyaaa-aaaaa-aaaba-cai&output=6c7su-kiaaa-aaaar-qaira-cai";
 
   return (
     <div className={className}>
       <div className="border border-orange-500 bg-orange-500/5 py-8 px-4 flex flex-col justify-center items-center rounded-xl text-center">
         <div className="mb-6 text-orange-500">
           <div className="font-semibold mb-2">
-            You don't have enough GLDT to process
+            You don't have enough GLDT to process.
           </div>
           <div>
             To reverse swap the{" "}
@@ -26,11 +23,17 @@ const InsufficientGLDTDisclaimer = ({ className }: { className?: string }) => {
             <span className="font-semibold">{totalSwapGLDT} GLDT</span>.
             <br />
             Your current balance is{" "}
-            <span className="font-semibold">{balanceGLDT} GLDT</span>
+            <span className="font-semibold">{balanceGLDT} GLDT.</span>
           </div>
         </div>
         <div>
-          <Button onClick={handleClickGetGLDT}>Get GLDT</Button>
+          <Link
+            to={linkIcpSwapPairICPGLDT}
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            <Button>Get GLDT</Button>
+          </Link>
         </div>
       </div>
     </div>