+++ client/gldt_swap_app/src/hooks/gld_nft/useFetchTransferFeeNft.tsx
@@ -2,6 +2,7 @@ import { useQuery, keepPreviousData } from "@tanstack/react-query";
 
 import { useAuth } from "@auth/index";
 import { divideBy1e8 } from "@utils/numbers";
+import { OGY_TX_FEE } from "@constants";
 
 export const useFetchTransferFeeNft = ({
   nftId,
@@ -23,7 +24,7 @@ export const useFetchTransferFeeNft = ({
     const result = (await actor.icrc7_transfer_fee(
       BigInt(nftId as bigint)
     )) as [bigint];
-    return divideBy1e8(result[0]) ?? undefined;
+    return divideBy1e8(result[0] + BigInt(OGY_TX_FEE)) ?? undefined;
   };
 
   return useQuery({