+++ client/gldt_swap_app/src/utils/numbers/index.ts
@@ -8,7 +8,7 @@ interface roundAndFormatLocaleParams {
     decimals?: number;
 }
 
-export const roundAndFormatLocale = ({number, locale="en-US", decimals=2}: roundAndFormatLocaleParams) => {
+export const roundAndFormatLocale = ({number, locale="en-US", decimals=3}: roundAndFormatLocaleParams) => {
     return Number(number.toFixed(decimals)).toLocaleString(locale);
 };